+++ src/keys/public_key.rs
@@ -423,268 +423,3 @@ impl GlobalPublicKey {
         Ok(polys)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::params::PvwParametersBuilder;
-    use rand::thread_rng;
-
-    /// Standard moduli suitable for PVW operations
-    fn test_moduli() -> Vec<u64> {
-        vec![0xffffee001u64, 0xffffc4001u64, 0x1ffffe0001u64]
-    }
-
-    /// Create PVW parameters for testing with moderate security settings
-    fn create_test_params() -> Arc<PvwParameters> {
-        PvwParametersBuilder::new()
-            .set_parties(5)
-            .set_dimension(1024)
-            .set_l(8)
-            .set_moduli(&test_moduli())
-            .set_secret_variance(1)
-            .set_error_bounds_u32(100, 200)
-            .build_arc()
-            .unwrap()
-    }
-
-    /// Create PVW parameters that satisfy the correctness condition
-    fn create_correct_test_params() -> Arc<PvwParameters> {
-        let moduli = test_moduli();
-
-        let (variance, bound1, bound2) =
-            PvwParameters::suggest_correct_parameters(5, 4, 8, &moduli).unwrap_or((1, 50, 100));
-
-        PvwParametersBuilder::new()
-            .set_parties(5)
-            .set_dimension(1024)
-            .set_l(8)
-            .set_moduli(&moduli)
-            .set_secret_variance(variance)
-            .set_error_bounds_u32(bound1, bound2)
-            .build_arc()
-            .unwrap()
-    }
-
-    #[test]
-    fn test_party_creation() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        let party = Party::new(0, &params, &mut rng).unwrap();
-        assert_eq!(party.index(), 0);
-        assert_eq!(party.secret_key().params.k, params.k);
-
-        // Test invalid index
-        let invalid_party = Party::new(params.n, &params, &mut rng);
-        assert!(invalid_party.is_err());
-    }
-
-    #[test]
-    fn test_party_creation_with_correct_parameters() {
-        let params = create_correct_test_params();
-        let mut rng = thread_rng();
-
-        let party = Party::new(0, &params, &mut rng).unwrap();
-        assert!(params.verify_correctness_condition());
-        assert_eq!(party.secret_key().params.k, params.k);
-    }
-
-    #[test]
-    fn test_public_key_generation() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        let party = Party::new(0, &params, &mut rng).unwrap();
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let pk = party.generate_public_key(&crs, &mut rng).unwrap();
-
-        assert_eq!(pk.dimension(), params.k);
-        assert!(pk.validate().is_ok());
-
-        // Test polynomial access
-        for i in 0..params.k {
-            let poly = pk.get_polynomial(i).unwrap();
-            assert!(Arc::ptr_eq(&poly.ctx, &params.context));
-        }
-    }
-
-    #[test]
-    fn test_global_public_key() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let global_pk = GlobalPublicKey::new(crs);
-
-        assert_eq!(global_pk.dimensions(), (params.n, params.k));
-        assert_eq!(global_pk.num_public_keys(), 0);
-        assert!(!global_pk.is_full());
-        assert!(global_pk.validate().is_ok());
-    }
-
-    #[test]
-    fn test_key_generation_workflow() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        // Create parties
-        let party_0 = Party::new(0, &params, &mut rng).unwrap();
-        let party_1 = Party::new(1, &params, &mut rng).unwrap();
-
-        // Create CRS and global public key
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let mut global_pk = GlobalPublicKey::new(crs);
-
-        // Generate and add public keys
-        global_pk
-            .generate_and_add_party(&party_0, &mut rng)
-            .unwrap();
-        global_pk
-            .generate_and_add_party(&party_1, &mut rng)
-            .unwrap();
-
-        assert_eq!(global_pk.num_public_keys(), 2);
-
-        // Test retrieving public keys
-        let retrieved_pk_0 = global_pk.get_public_key(0).unwrap();
-        let retrieved_pk_1 = global_pk.get_public_key(1).unwrap();
-
-        assert!(retrieved_pk_0.validate().is_ok());
-        assert!(retrieved_pk_1.validate().is_ok());
-
-        // Test polynomial access
-        let party_0_polys = global_pk.get_party_polynomials(0).unwrap();
-        assert_eq!(party_0_polys.len(), params.k);
-    }
-
-    #[test]
-    fn test_batch_key_generation() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        // Create multiple parties
-        let parties: Vec<Party> = (0..3)
-            .map(|i| Party::new(i, &params, &mut rng).unwrap())
-            .collect();
-
-        // Create CRS and global public key
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let mut global_pk = GlobalPublicKey::new(crs);
-
-        // Generate all keys at once
-        global_pk.generate_all_party_keys(&parties).unwrap();
-
-        assert_eq!(global_pk.num_public_keys(), 3);
-        assert!(!global_pk.is_full()); // 3 out of 5 parties
-
-        // Verify all keys are valid
-        for i in 0..3 {
-            let pk = global_pk.get_public_key(i).unwrap();
-            assert!(pk.validate().is_ok());
-        }
-    }
-
-    #[test]
-    fn test_secret_key_batch_generation() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        // Create secret keys directly
-        let secret_keys: Vec<SecretKey> = (0..2)
-            .map(|_| SecretKey::random(&params, &mut rng).unwrap())
-            .collect();
-
-        // Create CRS and global public key
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let mut global_pk = GlobalPublicKey::new(crs);
-
-        // Generate all keys from secret keys
-        global_pk.generate_all_keys(&secret_keys).unwrap();
-
-        assert_eq!(global_pk.num_public_keys(), 2);
-
-        // Verify generated keys
-        for i in 0..2 {
-            let pk = global_pk.get_public_key(i).unwrap();
-            assert!(pk.validate().is_ok());
-            assert_eq!(pk.dimension(), params.k);
-        }
-    }
-
-    #[test]
-    fn test_public_key_retrieval() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        let party = Party::new(0, &params, &mut rng).unwrap();
-        let crs = PvwCrs::new(&params, &mut rng).unwrap();
-        let mut global_pk = GlobalPublicKey::new(crs);
-
-        // Add a key
-        global_pk.generate_and_add_party(&party, &mut rng).unwrap();
-
-        // Test different retrieval methods
-        let pk = global_pk.get_public_key(0).unwrap();
-        let polys = global_pk.get_party_polynomials(0).unwrap();
-
-        assert_eq!(pk.dimension(), params.k);
-        assert_eq!(polys.len(), params.k);
-
-        // Test out of bounds access
-        assert!(global_pk.get_public_key(5).is_none());
-        assert!(global_pk.get_party_polynomials(5).is_err());
-    }
-
-    #[test]
-    fn test_dimension_validation() {
-        let params = create_test_params();
-        let mut rng = thread_rng();
-
-        // Create mismatched parameters for testing
-        let wrong_params = PvwParametersBuilder::new()
-            .set_parties(5)
-            .set_dimension(8) // Different k
-            .set_l(8)
-            .set_moduli(&test_moduli())
-            .set_secret_variance(1)
-            .set_error_bounds_u32(100, 200)
-            .build_arc()
-            .unwrap();
-
-        let party = Party::new(0, &params, &mut rng).unwrap();
-        let wrong_crs = PvwCrs::new(&wrong_params, &mut rng).unwrap();
-
-        // This should fail due to dimension mismatch
-        let result = party.generate_public_key(&wrong_crs, &mut rng);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_full_capacity() {
-        let small_params = PvwParametersBuilder::new()
-            .set_parties(2) // Only 2 parties
-            .set_dimension(2)
-            .set_l(8)
-            .set_moduli(&test_moduli())
-            .set_secret_variance(1)
-            .set_error_bounds_u32(100, 200)
-            .build_arc()
-            .unwrap();
-
-        let mut rng = thread_rng();
-        let crs = PvwCrs::new(&small_params, &mut rng).unwrap();
-        let mut global_pk = GlobalPublicKey::new(crs);
-
-        assert!(!global_pk.is_full());
-
-        // Add keys for both parties
-        for i in 0..2 {
-            let party = Party::new(i, &small_params, &mut rng).unwrap();
-            global_pk.generate_and_add_party(&party, &mut rng).unwrap();
-        }
-
-        assert!(global_pk.is_full());
-        assert_eq!(global_pk.num_public_keys(), 2);
-    }
-}
