+++ src/params/parameters.rs
@@ -602,422 +602,3 @@ impl PvwParameters {
         ))
     }
 }
-
-// Example usage with the specified moduli
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_pvw_parameters_with_custom_moduli() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        // Try different parameter combinations
-        let test_cases = [
-            (3, 8, 8),      // Small parameters
-            (10, 128, 16),  // Medium parameters
-            (50, 2048, 32), // Larger parameters
-        ];
-
-        for (n, k, l) in test_cases {
-            println!("\n=== Testing parameters: n={n}, k={k}, l={l} ===");
-
-            // Get suggested correct parameters
-            match PvwParameters::suggest_correct_parameters(n, k, l, &moduli) {
-                Ok((secret_variance, error_bound_1, error_bound_2)) => {
-                    println!(
-                        "Suggested parameters found: variance={secret_variance}, bound1={error_bound_1}, bound2={error_bound_2}"
-                    );
-
-                    // Create parameters with suggested bounds
-                    match PvwParameters::new_with_u32_bounds(
-                        n,
-                        k,
-                        l,
-                        &moduli,
-                        secret_variance,
-                        error_bound_1,
-                        error_bound_2,
-                    ) {
-                        Ok(params) => {
-                            println!(
-                                "Parameters created successfully with Delta = {}",
-                                params.delta()
-                            );
-
-                            // Verify they satisfy the correctness condition
-                            match params.verify_parameters() {
-                                Ok(true) => println!("✓ All parameter checks passed!"),
-                                Ok(false) => println!("✗ Parameter verification failed"),
-                                Err(e) => println!("✗ Verification error: {e}"),
-                            }
-                        }
-                        Err(e) => println!("✗ Failed to create parameters: {e}"),
-                    }
-                }
-                Err(e) => println!("✗ Could not find suitable parameters: {e}"),
-            }
-        }
-    }
-
-    #[test]
-    fn example_usage() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        // Method 1: Using builder pattern with custom parameters
-        let params1 = PvwParameters::builder()
-            .set_parties(5)
-            .set_dimension(128)
-            .set_l(16)
-            .set_moduli(&moduli)
-            .set_secret_variance(2)
-            .set_error_bounds_u32(500, 1000)
-            .build();
-
-        match params1 {
-            Ok(p) => {
-                println!("Method 1 - Parameters created with Delta = {}", p.delta());
-                if p.verify_correctness_condition() {
-                    println!("✓ Correctness condition satisfied");
-                } else {
-                    println!("✗ Correctness condition NOT satisfied");
-                }
-            }
-            Err(e) => println!("Method 1 failed: {e}"),
-        }
-
-        // Method 2: Using direct constructor
-        let params2 = PvwParameters::new_with_u32_bounds(
-            3,       // n (parties)
-            64,      // k (LWE dimension)
-            8,       // l (redundancy parameter)
-            &moduli, // specified moduli
-            1,       // secret_variance
-            200,     // error_bound_1
-            400,     // error_bound_2
-        );
-
-        match params2 {
-            Ok(p) => {
-                println!("Method 2 - Parameters created with Delta = {}", p.delta());
-                if p.verify_correctness_condition() {
-                    println!("✓ Correctness condition satisfied");
-                } else {
-                    println!("✗ Correctness condition NOT satisfied");
-                }
-            }
-            Err(e) => println!("Method 2 failed: {e}"),
-        }
-
-        // Method 3: Get suggested correct parameters first
-        if let Ok((variance, bound1, bound2)) =
-            PvwParameters::suggest_correct_parameters(5, 128, 16, &moduli)
-        {
-            let params3 =
-                PvwParameters::new_with_u32_bounds(5, 128, 16, &moduli, variance, bound1, bound2);
-
-            match params3 {
-                Ok(p) => {
-                    println!("Method 3 - Correct parameters created with suggested bounds");
-                    println!("Delta = {}", p.delta());
-                    assert!(
-                        p.verify_correctness_condition(),
-                        "Should satisfy correctness condition"
-                    );
-                }
-                Err(e) => println!("Method 3 failed: {e}"),
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod bigint_conversion_tests {
-    use super::*;
-    use num_traits::{One, Zero};
-
-    #[test]
-    fn test_bigints_to_poly_basic() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test case 1: Zero polynomial
-        let zero_coeffs = vec![BigInt::zero(); params.l];
-        let zero_poly = params
-            .bigints_to_poly(&zero_coeffs)
-            .expect("Failed to convert zero coefficients");
-
-        // Convert back to BigUint and verify
-        let recovered_zero: Vec<BigUint> = (&zero_poly).into();
-        for (i, coeff) in recovered_zero.iter().enumerate() {
-            assert_eq!(
-                *coeff,
-                BigUint::zero(),
-                "Zero coefficient {i} should remain zero"
-            );
-        }
-        println!("✓ Zero polynomial test passed");
-
-        // Test case 2: Simple positive values
-        let simple_coeffs: Vec<BigInt> = (1..=params.l).map(BigInt::from).collect();
-        let simple_poly = params
-            .bigints_to_poly(&simple_coeffs)
-            .expect("Failed to convert simple coefficients");
-
-        let recovered_simple: Vec<BigUint> = (&simple_poly).into();
-        for (i, (original, recovered)) in simple_coeffs
-            .iter()
-            .zip(recovered_simple.iter())
-            .enumerate()
-        {
-            let expected = original.to_biguint().unwrap();
-            assert_eq!(*recovered, expected, "Simple coefficient {i} mismatch");
-        }
-        println!("✓ Simple positive values test passed");
-
-        // Test case 3: Large values (within modulus range)
-        let delta = params.delta().clone();
-        let large_coeffs: Vec<BigInt> = (0..params.l)
-            .map(|i| BigInt::from(delta.clone()) * BigInt::from(i + 1))
-            .collect();
-
-        let large_poly = params
-            .bigints_to_poly(&large_coeffs)
-            .expect("Failed to convert large coefficients");
-
-        let recovered_large: Vec<BigUint> = (&large_poly).into();
-        for (i, (original, recovered)) in
-            large_coeffs.iter().zip(recovered_large.iter()).enumerate()
-        {
-            let expected_reduced = (original % BigInt::from(params.q_total()))
-                .to_biguint()
-                .unwrap();
-            assert_eq!(
-                *recovered, expected_reduced,
-                "Large coefficient {i} mismatch"
-            );
-        }
-        println!("✓ Large values test passed");
-    }
-
-    #[test]
-    fn test_bigints_to_poly_negative_values() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test negative values
-        let negative_coeffs: Vec<BigInt> = (1..=params.l).map(|i| -BigInt::from(i * 100)).collect();
-
-        let negative_poly = params
-            .bigints_to_poly(&negative_coeffs)
-            .expect("Failed to convert negative coefficients");
-
-        let recovered_negative: Vec<BigUint> = (&negative_poly).into();
-        let q_total = BigInt::from(params.q_total());
-
-        for (i, (original, recovered)) in negative_coeffs
-            .iter()
-            .zip(recovered_negative.iter())
-            .enumerate()
-        {
-            // For negative values, expect: (original % q_total + q_total) % q_total
-            let mut expected = original % &q_total;
-            if expected.is_negative() {
-                expected += &q_total;
-            }
-            let expected_biguint = expected.to_biguint().unwrap();
-
-            assert_eq!(
-                *recovered, expected_biguint,
-                "Negative coefficient {i} mismatch"
-            );
-        }
-        println!("✓ Negative values test passed");
-    }
-
-    #[test]
-    fn test_bigints_to_poly_round_trip() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test round-trip: BigInt → Poly → BigUint → BigInt
-        let original_coeffs: Vec<BigInt> = vec![
-            BigInt::from(42),
-            -BigInt::from(123),
-            BigInt::from(params.delta().clone()) / BigInt::from(2),
-            BigInt::zero(),
-            BigInt::one(),
-            -BigInt::one(),
-            BigInt::from(999999),
-            -BigInt::from(888888),
-        ];
-
-        // Convert to polynomial
-        let poly = params
-            .bigints_to_poly(&original_coeffs)
-            .expect("Failed to convert to polynomial");
-
-        // Convert back to BigUint
-        let recovered_biguints: Vec<BigUint> = (&poly).into();
-
-        // Convert back to BigInt and verify
-        let q_total_bigint = BigInt::from(params.q_total());
-        for (i, (original, recovered_biguint)) in original_coeffs
-            .iter()
-            .zip(recovered_biguints.iter())
-            .enumerate()
-        {
-            let recovered_bigint = BigInt::from(recovered_biguint.clone());
-
-            // Calculate expected value after modular reduction
-            let mut expected = original % &q_total_bigint;
-            if expected.is_negative() {
-                expected += &q_total_bigint;
-            }
-
-            assert_eq!(
-                recovered_bigint, expected,
-                "Round-trip failed for coefficient {i}: original={original}, expected={expected}, got={recovered_bigint}"
-            );
-        }
-        println!("✓ Round-trip conversion test passed");
-    }
-
-    #[test]
-    fn test_bigints_to_poly_gadget_polynomial() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test the gadget polynomial specifically
-        let gadget_poly = params
-            .gadget_polynomial()
-            .expect("Failed to create gadget polynomial");
-
-        // Ensure we're in PowerBasis representation for lifting
-        let mut poly_for_lifting = gadget_poly.clone();
-        poly_for_lifting.change_representation(Representation::PowerBasis);
-
-        // Use fhe.rs lift function to properly reconstruct coefficients
-        let gadget_coeffs: Vec<BigUint> = (&poly_for_lifting).into();
-
-        println!("Gadget coefficients after CRT lift:");
-        for (i, coeff) in gadget_coeffs.iter().enumerate() {
-            println!("  coeff[{i}] = {coeff}");
-        }
-
-        // Verify gadget structure: [1, Δ, Δ², ..., Δ^(ℓ-1)]
-        let mut expected_delta_power = BigUint::one();
-        for (i, coeff) in gadget_coeffs.iter().enumerate() {
-            assert_eq!(
-                *coeff, expected_delta_power,
-                "Gadget coefficient {i} should be Delta^{i} = {expected_delta_power}, got {coeff}"
-            );
-
-            if i < params.l - 1 {
-                expected_delta_power *= params.delta();
-            }
-        }
-        println!("✓ Gadget polynomial conversion test passed");
-    }
-
-    #[test]
-    fn test_bigints_to_poly_error_cases() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test wrong number of coefficients
-        let wrong_size_coeffs = vec![BigInt::from(1), BigInt::from(2)]; // Only 2 coeffs, need 8
-        let result = params.bigints_to_poly(&wrong_size_coeffs);
-        assert!(
-            result.is_err(),
-            "Should fail with wrong number of coefficients"
-        );
-        println!("✓ Wrong size error handling test passed");
-
-        // Test empty coefficients
-        let empty_coeffs: Vec<BigInt> = vec![];
-        let result = params.bigints_to_poly(&empty_coeffs);
-        assert!(result.is_err(), "Should fail with empty coefficients");
-        println!("✓ Empty coefficients error handling test passed");
-    }
-
-    #[test]
-    fn test_bigints_to_poly_performance() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Create test coefficients
-        let test_coeffs: Vec<BigInt> = (0..params.l)
-            .map(|i| BigInt::from(i * 12345 + 67890))
-            .collect();
-
-        // Time multiple conversions
-        let start_time = std::time::Instant::now();
-        let num_iterations = 100;
-
-        for _ in 0..num_iterations {
-            let _poly = params
-                .bigints_to_poly(&test_coeffs)
-                .expect("Failed to convert coefficients");
-        }
-
-        let elapsed = start_time.elapsed();
-        let avg_time = elapsed / num_iterations;
-
-        println!(
-            "✓ Performance test: {num_iterations} conversions in {elapsed:?}, avg {avg_time:?} per conversion"
-        );
-
-        // Should be reasonably fast (less than 1ms per conversion for small polynomials)
-        assert!(avg_time.as_millis() < 10, "Conversion should be fast");
-    }
-
-    #[test]
-    fn test_compare_with_fhe_direct_conversion() {
-        let moduli = [0xffffee001, 0xffffc4001, 0x1ffffe0001];
-
-        let params = PvwParameters::new_with_u32_bounds(3, 64, 8, &moduli, 1, 100, 200)
-            .expect("Failed to create parameters");
-
-        // Test values that should work with both approaches
-        let test_values_u64: Vec<u64> = (1..=params.l as u64).collect();
-        let test_values_i64: Vec<i64> = test_values_u64.iter().map(|&v| v as i64).collect();
-
-        // Method 1: Direct fhe.rs conversion from i64
-        let direct_poly = Poly::from_coefficients(&test_values_i64, &params.context)
-            .expect("Failed direct fhe.rs conversion");
-        let direct_recovered: Vec<BigUint> = (&direct_poly).into();
-
-        // Method 2: Our BigInt conversion
-        let bigint_coeffs: Vec<BigInt> = test_values_i64.iter().map(|&v| BigInt::from(v)).collect();
-        let bigint_poly = params
-            .bigints_to_poly(&bigint_coeffs)
-            .expect("Failed BigInt conversion");
-        let bigint_recovered: Vec<BigUint> = (&bigint_poly).into();
-
-        // Compare results
-        for (i, (direct, bigint)) in direct_recovered
-            .iter()
-            .zip(bigint_recovered.iter())
-            .enumerate()
-        {
-            assert_eq!(
-                *direct, *bigint,
-                "Coefficient {i} differs between direct and BigInt conversion"
-            );
-        }
-        println!("✓ Comparison with direct fhe.rs conversion test passed");
-    }
-}
